Proper cast to avoid underflow

+++ yt/server/node/data_node/job.cpp
@@ -568,7 +568,7 @@ private:
 
             YT_LOG_DEBUG("Enqueuing blocks for replication (Blocks: %v-%v)",
                 currentBlockIndex,
-                currentBlockIndex + writeBlocks.size() - 1);
+                currentBlockIndex + static_cast<int>(writeBlocks.size()) - 1);
 
             auto writeResult = writer->WriteBlocks(writeBlocks);
             if (!writeResult) {
